+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Table chart component
+- Map chart component
 
 ## [0.3.0] - 2023-03-14
 ### Added